+++ lib/helpers/force_class_searcher.dart
@@ -12,25 +12,25 @@ class ClassSearcher<T> {
       library.declarations.values.forEach((declaration) {
           if (declaration is ClassMirror) {
             ClassMirror cm = declaration;
-            _searchForClassOnMirror(values, cm);
+            _searchForClassOnMirror(values, cm, cm);
           }
       });
     });
     return values;
   }
   
-  void _searchForClassOnMirror(List<T> values, ClassMirror cm) {
+  void _searchForClassOnMirror(List<T> values, ClassMirror cm, ClassMirror original) {
     if (cm.hasReflectedType) {
         if (cm.reflectedType == T) {
-          InstanceMirror im = cm.newInstance(const Symbol(''), []);
+          InstanceMirror im = original.newInstance(const Symbol(''), []);
           var obj = im.reflectee;
           values.add(obj);
         } else {
           for (var si in cm.superinterfaces) {
-              _searchForClassOnMirror(values, si);
+              _searchForClassOnMirror(values, si, cm);
           }
           if (cm.superclass != null) {
-            _searchForClassOnMirror(values, cm.superclass);
+            _searchForClassOnMirror(values, cm.superclass, cm);
           }
         }
     }