+++ lib/helpers/force_metadata_value.dart
@@ -4,13 +4,16 @@ class MetaDataValue<T> {
   
   Symbol memberName;
   InstanceMirror instanceMirror;
+  MethodMirror _methodMirror;
   T object;
   
-  MetaDataValue(this.object, this.memberName, this.instanceMirror);
+  MetaDataValue(this.object, this._methodMirror, this.memberName, this.instanceMirror);
   
   InstanceMirror invoke(List positionalArguments) {
      return instanceMirror.invoke(memberName, positionalArguments);
   }
+  
+  List<ParameterMirror> get parameters => this._methodMirror.parameters;
 
   String toString() => "$object - $memberName";